cAER: correct frameRenderer declaration in visualizer.
git-svn-id: https://svn.code.sf.net/p/jaer/code/cAER/trunk@8014 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ modules/visualizer/visualizer.c
@@ -31,7 +31,7 @@ struct visualizer_state {
 	int16_t eventRendererSizeY;
 	size_t eventRendererSlowDown;
 	struct caer_statistics_state eventStatistics;
-	uint32_t *frameRenderer;
+	uint16_t *frameRenderer;
 	int32_t frameRendererSizeX;
 	int32_t frameRendererSizeY;
 	int32_t frameRendererPositionX;
